Add get/set test_unit_format function declarations to iperf_api.h
The functions were added in f2c4e2b5bd343e06635613e87463b4cabfe98b27 and
implicitly declared in iperf_api.c, but were never actually declared in
the api header file! External users of the api need access to the
declaration in the header file to use the function from the library.