Make -c and -B work similar with respect to "%" interface specifications.
IPv6 link-local address literals are accepted in both places.
Note that we might not work correctly in the presumably rare case
of a FQDN that resolves to an IPv6 link-local address.

"%" is always treated as a separator for a shortcut to --bind-dev,
for IPv4 and IPv6 literals as well as hostnames passed to -c and -B.